requirements.txt
flask>=2.0
flask-cors
sqlalchemy>=1.4
pandas>=1.3
pdfplumber>=0.7
scikit-learn>=1.0
python-dateutil
joblib
